requirements.txt
torch>=2.0
numpy>=1.26
Pillow>=9.0
matplotlib>=3.5
scipy>=1.10
